API improvements
* Get BatfishException out of IpSpace
* Inline MockIpSpace inside the only test that uses it
