+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/IpWildcard.java
@@ -1,5 +1,8 @@
 package org.batfish.datamodel;
 
+import static com.google.common.base.Preconditions.checkArgument;
+import static com.google.common.base.Preconditions.checkState;
+
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
 import java.io.Serializable;
@@ -8,7 +11,6 @@ import java.util.Objects;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
-import org.batfish.common.BatfishException;
 
 /** An IP wildcard consisting of a IP address and a wildcard (also expressed as an IP address) */
 @ParametersAreNonnullByDefault
@@ -24,18 +26,12 @@ public final class IpWildcard implements Serializable, Comparable<IpWildcard> {
   private static Ip parseAddress(String str) {
     if (str.contains(":")) {
       String[] parts = str.split(":");
-      if (parts.length != 2) {
-        throw new BatfishException("Invalid IpWildcard string: '" + str + "'");
-      } else {
-        return Ip.parse(parts[0]);
-      }
+      checkArgument(parts.length == 2, "Invalid IpWildcard string: '%s'");
+      return Ip.parse(parts[0]);
     } else if (str.contains("/")) {
       String[] parts = str.split("/");
-      if (parts.length != 2) {
-        throw new BatfishException("Invalid IpWildcard string: '" + str + "'");
-      } else {
-        return Ip.parse(parts[0]);
-      }
+      checkArgument(parts.length == 2, "Invalid IpWildcard string: '%s'");
+      return Ip.parse(parts[0]);
     } else {
       return Ip.parse(str);
     }
@@ -44,19 +40,13 @@ public final class IpWildcard implements Serializable, Comparable<IpWildcard> {
   private static Ip parseMask(String str) {
     if (str.contains(":")) {
       String[] parts = str.split(":");
-      if (parts.length != 2) {
-        throw new BatfishException("Invalid IpWildcard string: '" + str + "'");
-      } else {
-        return Ip.parse(parts[1]);
-      }
+      checkArgument(parts.length == 2, "Invalid IpWildcard string: '%s'");
+      return Ip.parse(parts[1]);
     } else if (str.contains("/")) {
       String[] parts = str.split("/");
-      if (parts.length != 2) {
-        throw new BatfishException("Invalid IpWildcard string: '" + str + "'");
-      } else {
-        int prefixLength = Integer.parseInt(parts[1]);
-        return Ip.numSubnetBitsToSubnetMask(prefixLength).inverted();
-      }
+      checkArgument(parts.length == 2, "Invalid IpWildcard string: '%s'");
+      int prefixLength = Integer.parseInt(parts[1]);
+      return Ip.numSubnetBitsToSubnetMask(prefixLength).inverted();
     } else {
       return Ip.ZERO;
     }
@@ -70,9 +60,7 @@ public final class IpWildcard implements Serializable, Comparable<IpWildcard> {
     // Canonicalize the address before passing it to parent, so that #equals works.
     _ip = Ip.create(address.asLong() & ~wildcardMask.asLong());
     _mask = wildcardMask;
-    if (!wildcardMask.valid()) {
-      throw new BatfishException("Invalid wildcard: " + wildcardMask);
-    }
+    checkArgument(wildcardMask.valid(), "Invalid wildcard: %s", wildcardMask);
   }
 
   public IpWildcard(Prefix prefix) {
@@ -180,11 +168,8 @@ public final class IpWildcard implements Serializable, Comparable<IpWildcard> {
   }
 
   public Prefix toPrefix() {
-    if (isPrefix()) {
-      return Prefix.create(_ip, _mask.inverted());
-    } else {
-      throw new BatfishException("Invalid wildcard format for conversion to prefix: " + _mask);
-    }
+    checkState(isPrefix(), "Invalid wildcard format for conversion to prefix: %s", _mask);
+    return Prefix.create(_ip, _mask.inverted());
   }
 
   @JsonValue
