+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/Prefix.java
@@ -13,7 +13,6 @@ import java.util.concurrent.ExecutionException;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
-import org.batfish.common.BatfishException;
 
 /** An IPv4 Prefix */
 @ParametersAreNonnullByDefault
@@ -51,15 +50,13 @@ public final class Prefix implements Comparable<Prefix>, Serializable {
   public static Prefix parse(@Nullable String text) {
     checkArgument(text != null, "Invalid IPv4 prefix %s", text);
     String[] parts = text.split("/");
-    if (parts.length != 2) {
-      throw new BatfishException("Invalid prefix string: \"" + text + "\"");
-    }
+    checkArgument(parts.length == 2, "Invalid prefix string: \"%s\"", text);
     Ip ip = Ip.parse(parts[0]);
     int prefixLength;
     try {
       prefixLength = Integer.parseInt(parts[1]);
     } catch (NumberFormatException e) {
-      throw new BatfishException("Invalid prefix length: \"" + parts[1] + "\"", e);
+      throw new IllegalArgumentException("Invalid prefix length: \"" + parts[1] + "\"", e);
     }
     return create(ip, prefixLength);
   }
@@ -72,7 +69,7 @@ public final class Prefix implements Comparable<Prefix>, Serializable {
   public static Optional<Prefix> tryParse(@Nonnull String text) {
     try {
       return Optional.of(parse(text));
-    } catch (IllegalArgumentException | BatfishException e) {
+    } catch (IllegalArgumentException e) {
       return Optional.empty();
     }
   }
