+++ projects/batfish-common-protocol/src/test/java/org/batfish/datamodel/ForwardingAnalysisImplTest.java
@@ -42,9 +42,11 @@ import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.ImmutableSortedMap;
 import com.google.common.collect.ImmutableSortedSet;
 import java.util.Map;
+import java.util.Objects;
 import java.util.Set;
 import java.util.SortedMap;
 import org.batfish.common.topology.TopologyUtil;
+import org.batfish.datamodel.visitors.GenericIpSpaceVisitor;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -1658,5 +1660,47 @@ public class ForwardingAnalysisImplTest {
             .get(v1.getName())
             .get(i1.getName())
             .containsIp(ip2, c1.getIpSpaces()));
+  }
+
+  private static class MockIpSpace extends IpSpace {
+
+    /** */
+    private static final long serialVersionUID = 1L;
+
+    private final int _num;
+
+    public MockIpSpace(int num) {
+      _num = num;
+    }
+
+    @Override
+    public <R> R accept(GenericIpSpaceVisitor<R> visitor) {
+      throw new UnsupportedOperationException("no implementation for generated method");
+    }
+
+    @Override
+    protected int compareSameClass(IpSpace o) {
+      return Integer.compare(_num, ((MockIpSpace) o)._num);
+    }
+
+    @Override
+    public boolean containsIp(Ip ip, Map<String, IpSpace> namedIpSpaces) {
+      throw new UnsupportedOperationException("no implementation for generated method");
+    }
+
+    @Override
+    protected boolean exprEquals(Object o) {
+      return _num == ((MockIpSpace) o)._num;
+    }
+
+    @Override
+    public int hashCode() {
+      return Objects.hash(_num);
+    }
+
+    @Override
+    public String toString() {
+      return String.format("TestIpSpace%d", _num);
+    }
   }
 }
